+++ packages/avatars-avataaars-sprites/src/options.ts
@@ -8,9 +8,10 @@ type Options = {
   hairColor?: Array<'auburn' | 'black' | 'blonde' | 'brown' | 'pastel' | 'platinum' | 'red' | 'gray'>;
   accessories?: Array<'kurt' | 'prescription01' | 'prescription02' | 'round' | 'sunglasses' | 'wayfarers'>;
   accessoriesChance?: number;
+  accessoriesColor?: Array<'black' | 'blue' | 'gray' | 'heather' | 'pastel' | 'pink' | 'red' | 'white'>;
   facialHair?: Array<'medium' | 'light' | 'majestic' | 'fancy' | 'magnum'>;
   facialHairChance?: number;
-  facialHairColor?: Array<'auburn' | 'black' | 'blonde' | 'brown' | 'platinum' | 'red'>;
+  facialHairColor?: Array<'auburn' | 'black' | 'blonde' | 'brown' | 'pastel' | 'platinum' | 'red' | 'gray'>;
   clothes?: Array<'blazer' | 'sweater' | 'shirt' | 'hoodie' | 'overall'>;
   clothesColor?: Array<'black' | 'blue' | 'gray' | 'heather' | 'pastel' | 'pink' | 'red' | 'white'>;
   eyes?: Array<
@@ -27,7 +28,7 @@ type Options = {
     | 'wink'
     | 'winkWacky'
   >;
-  eyebrow?: Array<'angry' | 'default' | 'flat' | 'raised' | 'sad' | 'unibrow' | 'up'>;
+  eyebrow?: Array<'angry' | 'default' | 'flat' | 'raised' | 'sad' | 'unibrow' | 'up' | 'frown'>;
   mouth?: Array<
     | 'concerned'
     | 'default'