+++ packages/avatars-avataaars-sprites/src/utils/getAccessoriesType.ts
@@ -1,38 +1,39 @@
-import Random from '@dicebear/avatars/lib/random';
-import Options from '../options';
+import type Random from '@dicebear/avatars/lib/random';
+import type Options from '../options';
 import getOption from './getOption';
+import { accessories } from '../paths';
 
-export default function(options: Options, random: Random) {
+export default function (options: Options, random: Random) {
   let accessoriesType = [];
 
   if (getOption('accessories', 'kurt', options)) {
-    accessoriesType.push('Kurt');
+    accessoriesType.push(accessories.kurt);
   }
 
   if (getOption('accessories', 'prescription01', options)) {
-    accessoriesType.push('Prescription01');
+    accessoriesType.push(accessories.prescription01);
   }
 
   if (getOption('accessories', 'prescription02', options)) {
-    accessoriesType.push('Prescription02');
+    accessoriesType.push(accessories.prescription02);
   }
 
   if (getOption('accessories', 'round', options)) {
-    accessoriesType.push('Round');
+    accessoriesType.push(accessories.round);
   }
 
   if (getOption('accessories', 'sunglasses', options)) {
-    accessoriesType.push('Sunglasses');
+    accessoriesType.push(accessories.sunglasses);
   }
 
   if (getOption('accessories', 'wayfarers', options)) {
-    accessoriesType.push('Wayfarers');
+    accessoriesType.push(accessories.wayfarers);
   }
 
   let pickedAccessoriesType = random.pickone(accessoriesType);
 
   if (false === random.bool(undefined !== options.accessoriesChance ? options.accessoriesChance : 10)) {
-    return 'Blank';
+    return undefined;
   }
 
   return pickedAccessoriesType;