+++ packages/avatars-avataaars-sprites/src/utils/getHairColor.ts
@@ -1,40 +1,41 @@
-import Random from '@dicebear/avatars/lib/random';
-import Options from '../options';
+import type Random from '@dicebear/avatars/lib/random';
+import type Options from '../options';
 import getOption from './getOption';
+import { hair } from '../colors';
 
-export default function(options: Options, random: Random) {
+export default function (options: Options, random: Random) {
   let hairColor = [];
 
   if (getOption('hairColor', 'auburn', options)) {
-    hairColor.push('Auburn');
+    hairColor.push(hair.auburn);
   }
 
   if (getOption('hairColor', 'black', options)) {
-    hairColor.push('Black');
+    hairColor.push(hair.black);
   }
 
   if (getOption('hairColor', 'blonde', options)) {
-    hairColor.push('Blonde', 'BlondeGolden');
+    hairColor.push(hair.blonde, hair.blondeGolden);
   }
 
   if (getOption('hairColor', 'brown', options)) {
-    hairColor.push('Brown', 'BrownDark');
+    hairColor.push(hair.brown, hair.brownDark);
   }
 
   if (getOption('hairColor', 'pastel', options)) {
-    hairColor.push('PastelPink');
+    hairColor.push(hair.pastelPink);
   }
 
   if (getOption('hairColor', 'platinum', options)) {
-    hairColor.push('Platinum');
+    hairColor.push(hair.platinum);
   }
 
   if (getOption('hairColor', 'red', options)) {
-    hairColor.push('Red');
+    hairColor.push(hair.red);
   }
 
   if (getOption('hairColor', 'gray', options)) {
-    hairColor.push('SilverGray');
+    hairColor.push(hair.silverGray);
   }
 
   return random.pickone(hairColor);