+++ packages/avatars-avataaars-sprites/src/utils/getNoseType.ts
@@ -1,0 +1,5 @@
+import { nose } from '../paths';
+
+export default function () {
+  return nose.default;
+}