+++ packages/avatars-avataaars-sprites/src/utils/getSkinType.ts
@@ -1,0 +1,5 @@
+import { skin } from '../paths';
+
+export default function () {
+  return skin.default;
+}