+++ packages/avatars-avataaars-sprites/src/utils/getTopType.ts
@@ -1,67 +1,79 @@
-import Random from '@dicebear/avatars/lib/random';
-import Options from '../options';
+import type Random from '@dicebear/avatars/lib/random';
+import type Options from '../options';
 import getOption from './getOption';
+import { top } from '../paths';
+import getHatColor from './getHatColor';
+import getHairColor from './getHairColor';
 
-export default function(options: Options, random: Random) {
+export default function (options: Options, random: Random) {
+  let hatColor = getHatColor(options, random);
+  let hairColor = getHairColor(options, random);
+
   let topType = [];
 
   if (getOption('top', 'longHair', options)) {
     topType.push(
-      'LongHairBigHair',
-      'LongHairBob',
-      'LongHairBun',
-      'LongHairCurly',
-      'LongHairCurvy',
-      'LongHairDreads',
-      'LongHairFrida',
-      'LongHairFro',
-      'LongHairFroBand',
-      'LongHairMiaWallace',
-      'LongHairNotTooLong',
-      'LongHairShavedSides',
-      'LongHairStraight',
-      'LongHairStraight2',
-      'LongHairStraightStrand'
+      () => top.bigHair(hairColor),
+      () => top.bob(hairColor),
+      () => top.bun(hairColor),
+      () => top.curly(hairColor),
+      () => top.curvy(hairColor),
+      () => top.dreads(hairColor),
+      () => top.frida(hairColor),
+      () => top.fro(hairColor),
+      () => top.froAndBand(hairColor),
+      () => top.miaWallace(hairColor),
+      () => top.longButNotTooLong(hairColor),
+      () => top.shavedSides(hairColor),
+      () => top.straight01(hairColor),
+      () => top.straight02(hairColor),
+      () => top.straightAndStrand(hairColor)
     );
   }
 
   if (getOption('top', 'shortHair', options)) {
     topType.push(
-      'ShortHairDreads01',
-      'ShortHairDreads02',
-      'ShortHairFrizzle',
-      'ShortHairShaggy',
-      'ShortHairShaggyMullet',
-      'ShortHairShortCurly',
-      'ShortHairShortFlat',
-      'ShortHairShortRound',
-      'ShortHairShortWaved',
-      'ShortHairSides',
-      'ShortHairTheCaesar',
-      'ShortHairTheCaesarSidePart'
+      () => top.dreads01(hairColor),
+      () => top.dreads02(hairColor),
+      () => top.frizzle(hairColor),
+      () => top.shaggy(hairColor),
+      () => top.shaggyMullet(hairColor),
+      () => top.shortCurly(hairColor),
+      () => top.shortFlat(hairColor),
+      () => top.shortRound(hairColor),
+      () => top.shortWaved(hairColor),
+      () => top.sides(hairColor),
+      () => top.theCaesar(hairColor),
+      () => top.theCaesarAndSidePart(hairColor)
     );
   }
 
   if (getOption('top', 'eyepatch', options)) {
-    topType.push('Eyepatch');
+    topType.push(() => top.eyepatch(hatColor));
   }
 
   if (getOption('top', 'hat', options)) {
-    topType.push('Hat', 'WinterHat1', 'WinterHat2', 'WinterHat3', 'WinterHat4');
+    topType.push(
+      () => top.hat(hatColor),
+      () => top.winterHat01(hatColor),
+      () => top.winterHat02(hatColor),
+      () => top.winterHat03(hatColor),
+      () => top.winterHat04(hatColor)
+    );
   }
 
   if (getOption('top', 'hijab', options)) {
-    topType.push('Hijab');
+    topType.push(() => top.hijab(hatColor));
   }
 
   if (getOption('top', 'turban', options)) {
-    topType.push('Turban');
+    topType.push(() => top.turban(hatColor));
   }
 
   let pickedTopType = random.pickone(topType);
 
   if (false === random.bool(undefined !== options.topChance ? options.topChance : 100)) {
-    return 'NoHair';
+    return undefined;
   }
 
   return pickedTopType;